MTL-1708 Built SP4 and SP3 images
This implements a matrix to build against any SLES base image and
publishes more specific Docker containing the distro information.

A basic tag will still be provided that contains just the Go
version, however this tag will only be published for whichever SLES
version is set in `mainSleVersion` in the `Jenkinsfile`. So
`csm-docker-sle-go:1.19` will be built on SP4, but a user could use
`csm-docker-sle-go:1.19-SLES15.3` to grab the SP3 base version.

Lastly the `Makefile` was modified to work outside of our Jenkins
environment in order to facilitate local development.