Prevent hadolint from deactivating
The cron trigger is not necessary for hadolint since it only lints the
`Dockerfile` and isn't doing any security scanning.

If this repo goes 60 days without activity, the presence of the cron
trigger will cause the hadolint workflow to be deactivated. This means
PRs made after 60 days of activity will not get linted.